requirements.txt
Django>=5.0
gunicorn
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
python-dotenv